Added timezone information back into aws returned snapshot StartTime (needed for correct 'snapshot_date_in_seconds' date conversion).
More explicitly, the problem with completely truncating the aws returned StartTime is that the time component also includes the timezone marker ("Z" for Zulu).  Truncating this information means that the following 'date' call interprets the returned StartTime as the servers' local time... which is incorrect.

Note, "date" can interpret the time string returned from AWS as it's in ISO format.